% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterWrappers.R
\name{ADEKF}
\alias{ADEKF}
\title{Filtering of state-of-charge and health.}
\usage{
ADEKF(
  I,
  V,
  Temp,
  Time,
  theta_0,
  ocv_0,
  SOC_0,
  dual = TRUE,
  P = NULL,
  Q = NULL,
  R = NULL,
  K = 1,
  trace = TRUE,
  trace_limit = 1e+06
)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{Temp}{The observed temperature.}

\item{Time}{The observed time.}

\item{theta_0}{The initial values of the parameter vector.}

\item{ocv_0}{The parameters used for the third order polynomial used to model the OCV.}

\item{SOC_0}{The initial value of the SOC.}

\item{dual}{TRUE/FALSE: Should it be a dual, or a single, EKF?}

\item{P}{The prior conditional variance of the state equation.}

\item{Q}{The variance of the state equation.}

\item{R}{The variance of the observation error.}

\item{K}{Number of RC branches in the EEC.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{trace_limit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Dual extended Kalman filter for determining the latent state-of-charge and health state-equation.
}
