% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterWrappers.R
\name{ADUKF}
\alias{ADUKF}
\title{Filtering of state-of-charge and health.}
\usage{
ADUKF(
  I,
  V,
  Temp,
  Time,
  ocv_0,
  theta_0,
  SOC_0 = 0,
  K = 1,
  sigma_0 = NULL,
  R_X = NULL,
  R_V = NULL,
  trace = TRUE,
  trace_limit = 1e+06
)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{Temp}{The observed temperature.}

\item{Time}{The observed time.}

\item{ocv_0}{The parameters used for the third order polynomial used to model the OCV.}

\item{theta_0}{The initial values of the parameter vector.}

\item{SOC_0}{The initial value of the SOC.}

\item{K}{Number of RC branches in the EEC.}

\item{sigma_0}{A vector of parameters used for the unscented transformation: (alpha, beta, kappa).}

\item{R_X}{The variance of the state-equation.}

\item{R_V}{The variance of the observation-equation.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{trace_limit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Dual extended Kalman filter for determining the latent state-of-charge and health state-equation.
}
