% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{barletts_test}
\alias{barletts_test}
\title{Barlett's test}
\usage{
barletts_test(n, sd_est, alpha_level)
}
\arguments{
\item{n}{A vector containing the counts of the 'k' groups.}

\item{sd_est}{The estimated standard deviation of the 'k' groups.}
}
\value{
A list of the elements need to calculate the Barlett's test and the associated p-value.
}
\description{
Performs Barlett's test for equal variance in 'k' groups.
}
