% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatteryPrognostics.R
\docType{package}
\name{BatteryPrognostics}
\alias{BatteryPrognostics}
\title{Battery Prognostics}
\description{
State-space model used for filtering and prediction of the state-of-charge and the state-of-health of Li-ion batteries by voltage given current and temperature.
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
