% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxilliaryFunctions.R
\name{compare_windows}
\alias{compare_windows}
\title{Compare windows}
\usage{
compare_windows(I1, I2, SOC1, SOC2, W, R, epsilon, delta, trace = TRUE,
  trace_limit = 10000, return_tibble = TRUE)
}
\arguments{
\item{I1}{Numeric vector of length 'T1'.}

\item{I2}{Numeric vector of length 'T2'.}

\item{SOC1}{Numeric vector of length 'T1'.}

\item{SOC2}{Numeric vector of length 'T2'.}

\item{W}{Scalar -- the window size.}

\item{R}{Scalar -- the maximum allowed number of place the two windows exceed epsilon.}

\item{epsilon}{Tolerance used when comparing windows of 'I1' and 'I2'.}

\item{delta}{Tolerance used when comparing the initial index of the window in 'SOC1' and 'SOC2'.}

\item{return_tibble}{TRUE/FALSE: Should a tibble be returned?}
}
\value{
Either a list of lists, or a \link{tibble}.
}
\description{
Compare all windows, with size W, of two vectors, 'I1' and 'I2', with the added requirement that starting index of the windows being compared of two other vectors 'SOC1' and 'SOC2' are within some tolerance 'delta'. The windows are compared by counting the number of indices where they differ more than some tolerance epsilon.
}
\details{
No default values are used for 'W', 'epsilon', or 'delta' as the choise of e.g. 'epsilon' and 'delta' are heavily dependent on the scale of the 'Is' and the 'SOCs'. Furthermore, the 'SOCs' are needed to reduce the amount of "relevant" information, otherwise the method will run out of memory quickly.
}
