% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxilliaryFunctions.R
\name{compare_windows_ga}
\alias{compare_windows_ga}
\title{Compare windows using genetic algorithm}
\usage{
compare_windows_ga(
  I1,
  I2,
  V1,
  V2,
  Temp1,
  Temp2,
  MutationWindow,
  RestrictTemperature,
  RestrictVoltage,
  Wmin,
  Wmax,
  Imin,
  N_evolution = 2e+06,
  N_keep = 100,
  trace = TRUE,
  trace_limit = 10000,
  return_tibble = TRUE
)
}
\arguments{
\item{I1}{Numeric vector of length 'T1'.}

\item{I2}{Numeric vector of length 'T2'.}

\item{V1}{Numeric vector of length 'T1'.}

\item{V2}{Numeric vector of length 'T2'.}

\item{Temp1}{Numeric vector of length 'T1'.}

\item{Temp2}{Numeric vector of length 'T2'.}

\item{MutationWindow}{Scalar: The size of the mutation window.}

\item{RestrictTemperature}{Numeric: The maximal allowed initial difference between the temperatures.}

\item{RestrictVoltage}{Numeric: The maximal allowed difference between the beginning and end between the voltages.}

\item{Wmin}{Scalar: Minimum the window size.}

\item{Wmax}{Scalar: Maximum the window size.}

\item{Imin}{Scalar: Minumum allowed starting current (A).}

\item{N_evolution}{Scalar: The number of evolutions of the GA.}

\item{N_keep}{Scalar: The number of individuals to keep form the GA.}

\item{trace}{TRUE/FLASE: Show trace?}

\item{trace_limit}{Scalar: Limit the trace shown to 'trace_limit'.}

\item{return_tibble}{TRUE/FALSE: Should a tibble be returned?}
}
\value{
Either a list of lists, or a \link{tibble}.
}
\description{
Compare windows of two vectors 'I1' and 'I2' using variable window sizes in a multi-objective GA.
}
\details{
No default values are used for 'W', 'epsilon', or 'delta' as the choise of e.g. 'epsilon' and 'delta' are heavily dependent on the scale of the 'Is' and the 'SOCs'. Furthermore, the 'SOCs' are needed to reduce the amount of "relevant" information, otherwise the method will run out of memory quickly.
}
