% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correct_time}
\alias{correct_time}
\title{Corrects time}
\usage{
correct_time(I, V, T)
}
\arguments{
\item{I}{A vector containing the current.}

\item{V}{A vector containing the voltage.}

\item{T}{A vector containing the time.}
}
\value{
A matrix containing the corrected time, and interpolated current and voltage.
}
\description{
Correction of time in the case of inconsistant time-step sizes.
}
