% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEKFwrappers.R
\name{EKF_parameters}
\alias{EKF_parameters}
\title{Filtering of state-of-charge and health.}
\usage{
EKF_parameters(I, V, theta_0, SOC_0, Q_max, eta, P = NULL, dt = 1, K = 2,
  W = 100, trace = TRUE, traceLimit = 1e+06)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{theta_0}{The initial values of the parameter vector.}

\item{SOC_0}{The initial value of the SOC.}

\item{Q_max}{The maximum value of the capacity.}

\item{eta}{The capacity efficiency.}

\item{P}{The prior conditional variance of the state equation.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}

\item{W}{Size of the windows used to estimate the R and Q.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{traceLimit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Dual extended Kalman filter for determining the latent state-of-charge and health state-equation.
}
