% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{estimate_parameters_resistance}
\alias{estimate_parameters_resistance}
\title{Estimate resistance parameters}
\usage{
estimate_parameters_resistance(R, SOC, symmetric = FALSE)
}
\arguments{
\item{R}{The extracted battery resistance.}

\item{SOC}{The corresponding SOC of the extracted battery resistance.}

\item{symmetric}{TRUE/FALSE: Is the model symmetric? That is, can we assume 'b == c'?}
}
\value{
The mean parameters and standard deviations of the resistance model.
}
\description{
Estimates the parameters of the resistance model.
}
\details{
The resistance model assumes that 'log(R) = a + b log(SOC) + c log(1 - SOC) + epsilon', 
where epsilon follows a normal distribution with mean zero and variance sigma^2.
}
