% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{estimate_parameters_resistance_all}
\alias{estimate_parameters_resistance_all}
\title{Estimate resistance parameters}
\usage{
estimate_parameters_resistance_all(R, W, formula_soc, formula_current, data,
  number_of_clusters = 4, alpha_level = 0.05)
}
\arguments{
\item{R}{The extracted resistance -- either a character string or a vector.}

\item{W}{The time each resistance value belongs to -- either a character string or a vector.}

\item{formula_soc}{A formula for the SOC component.}

\item{formula_current}{A formula for the Current component.}

\item{data}{The data need for the components.}

\item{number_of_clusters}{The number of clusters passed to 'doParallel' and 'foreach'.}

\item{alpha_level}{The significance level used for Barlett's and LR tests.}
}
\value{
A matrix with a row for every week containing the mean parameters and standard deviations of the resistance model.
}
\description{
Estimates the parameters of the resistance model for every level of 'W'.
}
