% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{estimate_parameters_resistance_formula}
\alias{estimate_parameters_resistance_formula}
\title{Estimate resistance parameters}
\usage{
estimate_parameters_resistance_formula(R, formula_soc, formula_current, data)
}
\arguments{
\item{R}{The extracted battery resistance.}

\item{formula_soc}{A formula for the SOC component.}

\item{formula_current}{A formula for the Current component.}

\item{data}{The data need for the components.}
}
\value{
The mean parameters and standard deviations of the resistance model.
}
\description{
Estimates the parameters of the resistance model.
}
