% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{estimate_parameters_resistance_weeks}
\alias{estimate_parameters_resistance_weeks}
\title{Estimate resistance parameters for all weeks}
\usage{
estimate_parameters_resistance_weeks(R, SOC, W, number_of_clusters = 4,
  symmetric = FALSE, alpha_level = 0.05)
}
\arguments{
\item{R}{The extracted battery resistance.}

\item{SOC}{The corresponding SOC of the extracted battery resistance.}

\item{W}{The corresponding week of the extracted battery resistance.}

\item{number_of_clusters}{The number of clusters passed to 'doParallel' and 'foreach'.}

\item{symmetric}{TRUE/FALSE: Is the model symmetric? That is, can we assume 'b == c'?}

\item{alpha_level}{The significance level used for Barlett's and LR tests.}
}
\value{
A matrix with a row for every week containing the mean parameters and standard deviations of the resistance model.
}
\description{
Estimates the parameters of the resistance model for every week on record.
}
\details{
The resistance model assumes that 'log(R) = a + b log(SOC) + c log(1 - SOC) + epsilon', 
where epsilon follows a normal distribution with mean zero and variance sigma^2.
}
