% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{estimate_parameters_resistance_weeks_tibble}
\alias{estimate_parameters_resistance_weeks_tibble}
\title{Estimate resistance parameters for all weeks}
\usage{
estimate_parameters_resistance_weeks_tibble(
  R,
  SOC,
  W,
  estimated_parameters = NULL,
  number_of_clusters = 4,
  symmetric = FALSE
)
}
\arguments{
\item{R}{The extracted battery resistance.}

\item{SOC}{The corresponding SOC of the extracted battery resistance.}

\item{W}{The corresponding week of the extracted battery resistance.}

\item{estimated_parameters}{A matrix of estimated parameters containing one row for every week.}

\item{number_of_clusters}{The number of clusters passed to 'doParallel' and 'foreach'.}

\item{symmetric}{TRUE/FALSE: Is the model symmetric? That is, can we assume 'b == c'?}
}
\value{
A tibble containing the mean parameters, standard deviations of the resistance model, expected log-resistance, and the minimum possible expected log-resistance.
}
\description{
Estimates the parameters of the resistance model for every week on record. The estimated parameters are stored in a tibble for ease of use.
}
\details{
The resistance model assumes that 'log(R) = a + b log(SOC) + c log(1 - SOC) + epsilon', 
where epsilon follows a normal distribution with mean zero and variance sigma^2.
}
