% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{exact_week_probablity}
\alias{exact_week_probablity}
\title{Week probability}
\usage{
exact_week_probablity(R, SOC, resistance_parameters)
}
\arguments{
\item{R}{A new resistance.}

\item{SOC}{A new state-of-charge.}

\item{resistance_parameters}{A matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}
}
\value{
A numeric vector of probabilities.
}
\description{
The probability of each week given a resistance and SOC.
}
