% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{expected_residual_life}
\alias{expected_residual_life}
\title{Expected mean residual life}
\usage{
expected_residual_life(post_w)
}
\arguments{
\item{probability_w}{An object of class 'post_w', created with either the \link{exact_week_probability} or \link{week_probability}.}
}
\value{
Numeric.
}
\description{
Calculates the expected mean residual life of a 'post_w' class objcet.
}
