% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxilliaryFunctions.R
\name{extract_resistance}
\alias{extract_resistance}
\title{Extract resistance}
\usage{
extract_resistance(I, V, T_s = NULL, SOC = NULL, epsilon = 0.001,
  Q_max = 2.56, eta = 1, SOC_0 = 0, convert_to_tibble = FALSE)
}
\arguments{
\item{I}{Vector containing the current.}

\item{V}{Vector containing the voltage.}

\item{T_s}{Vector containing the time step sizes (in seconds). If 'NULL' it is set as 'seq_along(I)'.}

\item{SOC}{Vector containing the SOC. If 'NULL' it is estimated by Columb counting, using the parameter 'SOC_0' as a starting point.}

\item{epsilon}{The lower limit used for identifying change in current. Default is '1e-3'.}

\item{Q_max}{The maximal capacity of the battery cell.}

\item{eta}{The Coulombic effeciency.}

\item{SOC_0}{The initial value of the state-of-charge.}

\item{convert_to_tibble}{Organise in \link{tibble}.}
}
\value{
A list containing the extracted resistance and their time scales extracted from the profile.
}
\description{
Extracts the resistance from a dynamic current/voltage profile.
}
