% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCKLookupTable.R
\name{extractLookupTables}
\alias{extractLookupTables}
\title{Extract look-up tables.}
\usage{
extractLookupTables(current, voltage, time_s, t_k, SOCValues = seq(5, 95, 5),
  IValues = c(-10, -7.5, -5, -2.5, -1.25, -0.25, 0.25, 1.25, 2.5, 5, 7.5, 10),
  h = 1)
}
\arguments{
\item{current}{A charge/discharge cycling current profile. Note: assumes that negative current implies discharging of the battery.}

\item{voltage}{A charge/discharge cycling voltage profile.}

\item{time_s}{The sampling time.}

\item{t_k}{The time intervals for each of the RC-units.}

\item{SOCValues}{A vector of possible 'SOC' values.}

\item{IValues}{A vector of possible 'Current' values.}

\item{h}{R0 reduction fraction. NOTE: THIS SHOULD NOT BE NECESSARY AND, THEREFORE, SET TO 1.}
}
\value{
A list of look-up table matrices.
}
\description{
Extracts RC-K parameter look-up tables from simple charge/discharge cycling profile.
}
