% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULEstimation.R
\name{feature_vars}
\alias{feature_vars}
\title{Estimate feature AR models}
\usage{
feature_vars(data, l = 1)
}
\arguments{
\item{data}{A data-set with each column containing the features to be modelled.}

\item{l}{The lag of the AR feature models (NOTE: predict and forward only works with '\code{l = 1}').}
}
\value{
A list of AR models.
}
\description{
Wrapper for esay estimation of feature AR models in RUL estimation.
}
