% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULEstimation.R
\name{forward}
\alias{forward}
\title{Forward simulation}
\usage{
forward(object, ...)
}
\arguments{
\item{object}{An object to simulate forward in time.}

\item{...}{Additional arguments needed in forward simulation.}
}
\description{
\code{forward} is a generic function for the look-ahead simulation of various stochastic process models. This function will invoke a particular method which will depend on the \link{class} of the first argument.
}
