% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMExtensions.R
\name{gam_parameter_fit}
\alias{gam_parameter_fit}
\title{GAM parameter wrap}
\usage{
gam_parameter_fit(resistance_parameters)
}
\arguments{
\item{resistance_parameters}{A list containing a vector of weeks and a matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}
}
\value{
A list of GAM fits.
}
\description{
A wrapping function creating a list of GAM fits of a 'resistance_parameters'-object.
}
