% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{mean_residual_life}
\alias{mean_residual_life}
\title{Mean residual life}
\usage{
mean_residual_life(w, probability_w)
}
\arguments{
\item{w}{The current age of the battery.}

\item{probability_w}{An object of class 'post_w', created with either the \link{exact_week_probability} or \link{week_probability}.}
}
\value{
Numeric.
}
\description{
Calculates the mean residual life, given the current age of the battery, and a 'post_w' class objcet.
}
