% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParticleFilter.R
\name{non_linear_particle_filter}
\alias{non_linear_particle_filter}
\title{Filtering of state-of-charge and health.}
\usage{
non_linear_particle_filter(
  I,
  V,
  Temperature,
  Time,
  SOC_0,
  K = 2,
  N = 10000,
  trace = 100,
  ...
)
}
\arguments{
\item{Temperature}{The observed temperature profile.}

\item{Time}{The time increments.}

\item{SOC_0}{The initial state-of-charge.}

\item{K}{The number of RC branches used in the state-space model.}

\item{N}{The number of particles.}

\item{trace}{TRUE/FALSE: Show trace? If 'numeric' provided, then the trace is shown every 'trace' elements.}

\item{...}{Additional arguments including initialisation of parameters.}

\item{Current}{The observed current profile.}

\item{Voltage}{The observed voltage profile.}
}
\value{
List of estimated parameters, predicted voltage, and estimated SOC.
}
\description{
Particle filter for determining the latent state-of-charge, and non-linear optimisation for determining state-of-health.
}
