% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxilliaryFunctions.R
\name{plot_extracted_resistance}
\alias{plot_extracted_resistance}
\title{2d-density plot of the extracted resistance}
\usage{
plot_extracted_resistance(ext_res_tibble, restrict_dim_y = NULL,
  include_marginals = FALSE, dims = c("SOC", "R"), facet_variable = NULL,
  plot_type = "hex", n_bins = NULL)
}
\arguments{
\item{ext_res_tibble}{\link{tibble} made with the \link{extract_resistance}-function.}

\item{restrict_dim_y}{Quantiles used to restrict the y-axis. See details.}

\item{include_marginals}{TRUE/FALSE: Should marginal plots be included? Note: requires 'ggExtra'.}

\item{dims}{The number of dimensions to be plotted.}

\item{facet_variable}{String passed to 'facet_wrap'.}

\item{plot_type}{Character of the type of plot. Takes the values 'density', 'hex', and 'bin'.}

\item{n_bins}{The number of bins passed to the 'hex' and 'bin' plot types.}
}
\value{
ggplot-object.
}
\description{
A wrapping function calling ggplot with 'stat_density_2d' and 'ggMarginal' (from the 'ggExtra'-package).
}
\details{
If 'restrict_dim_y' has length 1 it is taken as a restriction on the upper quantile. If it is 'NULL' the y-axis is not restricted.
}
