% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{posterior_resistance}
\alias{posterior_resistance}
\title{Predict resistance}
\usage{
posterior_resistance(
  post_pred_var,
  resistance_parameters,
  SOC = NULL,
  SOC_ahead = NULL,
  R = NULL,
  W = NULL,
  trace = FALSE,
  trace_limit = 1000,
  equal_variance = FALSE,
  return_tibble = FALSE,
  ...
)
}
\arguments{
\item{post_pred_var}{An object of class ''.}

\item{resistance_parameters}{An object of class 'resistance_parameters'.}

\item{SOC}{A single value, an object of class 'histogram', an object of class '', or a 'function'.}

\item{R}{The extracted battery resistance.}

\item{W}{The corresponding week of the extracted battery resistance.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{trace_limit}{Limits the amount the trace is printed.}

\item{equal_variance}{TRUE/FALSE: Is the variance assumed to be equal.}

\item{return_tibble}{TRUE/FALSE: Should a tibble be returned? If 'FALSE', a matrix is returned.}

\item{...}{Additional information passed to 'SOC'.}
}
\value{
A matrix, or tibble, containing predictions of every simulation for every week.
}
\description{
The posterior of the battery resistance
}
