% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{posterior_soc_probability}
\alias{posterior_soc_probability}
\title{SOC probability}
\usage{
posterior_soc_probability(R, W, resistance_parameters, burn_in, n_sims, trace,
  traceLimit)
}
\arguments{
\item{R}{The extracted battery resistance.}

\item{W}{The corresponding week of the extracted battery resistance.}

\item{resistance_parameters}{A list containing the weeks and a matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}

\item{burn_in}{The number of initial samples removed as burn-in.}

\item{n_sims}{The number of samples generated by the MH.}

\item{trace}{TRUE/FALSE: Should a trace be shown?}

\item{traceLimit}{Numeric: How often should the trace be shown?}
}
\value{
A numeric vector of probabilities.
}
\description{
The marginal probability of SOC given a resistance and week.
}
