% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARExtensions.R
\name{predict_posterior_VAR}
\alias{predict_posterior_VAR}
\title{Predict posterior of VAR}
\usage{
predict_posterior_VAR(object, n_ahead = 1, n_sims = 10000)
}
\arguments{
\item{object}{A '\link{varest}' object.}

\item{n_ahead}{The number of forward predictions.}

\item{n_sims}{The number of simulations.}
}
\description{
Simulates observations from the posterior predictive distribution of a VAR. In particular, for an object of the class '\link{varest}' from the \link{vars} package.
}
