% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{resistance_expected}
\alias{resistance_expected}
\title{The expected log-resistance.}
\usage{
resistance_expected(beta_w, SOC)
}
\arguments{
\item{beta_w}{The mean parameters of the resistance.}

\item{SOC}{A vector of the SOC.}
}
\value{
Numeric
}
\description{
The expected log-resistance.
}
