% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{resistance_given_week}
\alias{resistance_given_week}
\title{Sample resistance}
\usage{
resistance_given_week(resistance_parameters, n_sims = 10000, burn_in = 1000,
  return_tibble = TRUE, trace = TRUE, traceLimit = 1000)
}
\arguments{
\item{resistance_parameters}{A list containing the weeks and a matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}

\item{n_sims}{The number of samples generated by the MH.}

\item{burn_in}{The number of initial samples removed as burn-in.}

\item{return_tibble}{TRUE/FALSE: Should a tibble be returned.}

\item{trace}{TRUE/FALSE: Should a trace be shown?}

\item{traceLimit}{Numeric. How often should the trace be shown?}
}
\description{
MH-sampling from the posterior resistance distribution dependent only on the week.
}
