% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMExtensions.R
\name{resistance_parameters_interpolate}
\alias{resistance_parameters_interpolate}
\title{Interpolate 'resistance_parameters'}
\usage{
resistance_parameters_interpolate(resistance_parameters, smooth = FALSE,
  limits = NULL)
}
\arguments{
\item{resistance_parameters}{A list containing a vector of weeks and a matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}

\item{smooth}{TRUE/FALSE: Should the fitted 'gam_list' be used to smooth the estimated parameters.}

\item{limits}{A vector of one or two elements indicating the limits of the 'W'.}
}
\value{
An interpolated object of class 'resistance_parameters'.
}
\description{
Interpolates parameters of missing weeks using GAM.
}
