% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{resistance_parameters_list}
\alias{resistance_parameters_list}
\title{Resistance parameters list}
\usage{
resistance_parameters_list(...)
}
\arguments{
\item{...}{A series of \link{resistance_parameters} objects.}
}
\value{
A list of class 'resistance_parameters_list'.
}
\description{
A list wrap function for \link{resistance_parameters} class objects.
}
