% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{resistance_sd}
\alias{resistance_sd}
\title{Standard deviation of log-resistance}
\usage{
resistance_sd(beta_w, R, SOC)
}
\arguments{
\item{beta_w}{The mean parameters of the resistance.}

\item{R}{The extracted battery resistance.}

\item{SOC}{The corresponding SOC of the extracted battery resistance.}
}
\value{
Numeric
}
\description{
Standard deviation of log-resistance
}
