% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RULEstimation.R
\name{rul}
\alias{rul}
\title{Simulating Remaining Useful Life}
\usage{
rul(
  X,
  y,
  health_model,
  n_ahead = 50,
  n_simulations = 100,
  l = 1,
  eol = 0.8,
  y0 = NULL,
  return_full = FALSE,
  ...
)
}
\arguments{
\item{X}{A matrix of features.}

\item{y}{A vector of health metrics.}

\item{health_model}{A function modelling the relationship between features and health metrics.}

\item{n_ahead}{Amount of time simulated forward.}

\item{n_simulations}{The number of simulations used in the forward simulation.}

\item{l}{The lag used in feature modelling.}

\item{eol}{A numeric representing the end-of-life criterion.}

\item{y0}{An initial value of the health metric used to calculate the end-of-life. If '\code{NULL}' (default) it uses the first value of '\code{y}'.}

\item{return_full}{\code{TRUE/FALSE}: Should all estimated and simulated objects be returned?}

\item{...}{A list of parameters passed to the \code{health_model} function.}
}
\value{
A list containing the following four elements are returned:
}
\description{
The set-up, simulation, and prediction of remaining useful life (RUL) based on a feature abstraction approach.
}
