% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxilliaryFunctions.R
\name{sample_histogram}
\alias{sample_histogram}
\title{Sample from histogram}
\usage{
sample_histogram(x, size = 1)
}
\arguments{
\item{x}{An object of class '\link{histogram}'.}

\item{size}{The number of samples to draw.}
}
\value{
Numeric vector.
}
\description{
Sample from the bins of a histogram.
}
