% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{sd_residual_life}
\alias{sd_residual_life}
\title{Standard deviation of mean residual life}
\usage{
sd_residual_life(post_w)
}
\arguments{
\item{probability_w}{An object of class 'post_w', created with either the \link{exact_week_probability} or \link{week_probability}.}
}
\value{
Numeric.
}
\description{
Calculates the standard deviation of the mean residual life of a 'post_w' class objcet.
}
