% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SOC}
\alias{SOC}
\title{State-of-charge filtering}
\usage{
SOC(I, V, trace)
}
\arguments{
\item{I}{A current pR0file.}

\item{V}{The observed voltage.}

\item{trace}{TRUE/FALSE: Show trace?}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Kalman filter for determining the latent state-of-charge state-equation.
}
