% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SODEFK}
\alias{SODEFK}
\title{Filtering of state-of-charge and health.}
\usage{
SODEFK(I, V, theta_0, SOC_0, C_max, dt, K, trace, traceLimit)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{theta_0}{The initial values of the parameter vector.}

\item{SOC_0}{The initial value of the SOC.}

\item{C_max}{The maximum value of the capacity.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{traceLimit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Dual extended Kalman filter for determining the latent state-of-charge and health state-equation.
}
