% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SOEFK}
\alias{SOEFK}
\title{Filtering of state-of-charge and health.}
\usage{
SOEFK(I, IF, V, parameters_0, SOC_0, dt, K, trace, traceLimit)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{parameters_0}{The initial parameter values.}

\item{SOC_0}{The initial value of the SOC.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{traceLimit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Extended Kalman filter for determining the latent state-of-charge and health state-equation.
}
