% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFKOptimisation.R
\name{SOEFK_optimisation}
\alias{SOEFK_optimisation}
\title{Optimise EFK}
\usage{
SOEFK_optimisation(I, V, SOC_0 = 0, dt = 1, K = 2)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{SOC_0}{The initial value of the SOC.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}
}
\value{
A list of the optimal parameter values.
}
\description{
Parameter optimisation of the SO-EFK function.
}
