% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{symmetry_test}
\alias{symmetry_test}
\title{Symmetry's test}
\usage{
symmetry_test(beta_w, R, SOC, W)
}
\arguments{
\item{beta_w}{A matrix of the estimated parameters in the non-symmetric model.}
}
\value{
A p-value for each week in the provided matrix.
}
\description{
Performs likelihood ratio test for symmertry.
}
