% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEKFwrappers.R
\name{UKFParameter}
\alias{UKFParameter}
\title{Filtering of parameters.}
\usage{
UKFParameter(I, V, theta_0, P_0, SOC_0, C_max, eta, dt = 1, K = 2,
  trace = TRUE, traceLimit = 1e+06)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{theta_0}{The initial values of the parameter vector.}

\item{P_0}{The prior conditional variance of the state equation.}

\item{SOC_0}{The initial value of the SOC.}

\item{C_max}{The maximum value of the capacity.}

\item{eta}{The capacity efficiency.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{traceLimit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Unscented Kalman filter for determining the latent parameters.
}
