% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEKFwrappers.R
\name{UKS_parameters}
\alias{UKS_parameters}
\title{Smoothing of parameters.}
\usage{
UKS_parameters(I, V, theta_0, P_0, SOC_0, C_max, eta = 1, sigma_0 = c(0.1,
  2, 10), lag = 5, M = 10, dt = 1, K = 2, R_0 = NULL, Q_0 = NULL,
  trace = TRUE, traceLimit = 1e+06)
}
\arguments{
\item{I}{A current profile.}

\item{V}{The observed voltage.}

\item{theta_0}{The initial values of the parameter vector.}

\item{P_0}{The prior conditional variance of the state equation.}

\item{SOC_0}{The initial value of the SOC.}

\item{C_max}{The maximum value of the capacity.}

\item{eta}{The capacity efficiency.}

\item{sigma_0}{A vector containing the three sigma point parameters.}

\item{lag}{The time 'lag' used for parameter smoothing.}

\item{M}{Half the window size used for weighted moving average of parameters.}

\item{dt}{Step-size.}

\item{K}{Number of RC branches in the EEC.}

\item{trace}{TRUE/FALSE: Show trace?}

\item{traceLimit}{Used to limit trace.}
}
\value{
A list containing the voltage, the filtered voltage, the SOC, the polarisation voltage, and the state variance.
}
\description{
Unscented Kalman filter and smoother for determining the latent parameters.
}
