% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARExtensions.R
\name{VAR_resistance_parameters}
\alias{VAR_resistance_parameters}
\title{Wrapper for VAR}
\usage{
VAR_resistance_parameters(resistance_parameters, ...)
}
\arguments{
\item{resistance_parameters}{A list containing a vector of weeks and a matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}

\item{...}{Arguments passed to the \link{VAR}-function.}
}
\value{
A \link{varest}-object.
}
\description{
A wrapper for the VAR function for the class 'resistance_parameters'. The function also interpolates weeks not found in the provided 'resistance_parameters'-object.
}
