% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactResistanceEstimation.R
\name{week_probability}
\alias{week_probability}
\title{Week probability}
\usage{
week_probability(R, resistance_parameters, SOC = dunif, ...)
}
\arguments{
\item{R}{A new resistance.}

\item{resistance_parameters}{A matrix of estimated parameters constructed using \link{estimate_parameters_resistance_weeks}.}

\item{SOC}{Distribution of the SOC.}

\item{...}{Additional arguments passed to the SOC distribution, If 'SOC' is a function.}
}
\value{
A numeric vector of probabilities.
}
\description{
The probability of each week given a resistance and SOC.
}
