% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactResistanceEstimation.R
\name{weighted_quantile_rul}
\alias{weighted_quantile_rul}
\title{Weigthed quantiles of remaining useful life.}
\usage{
weighted_quantile_rul(probabilities_w, ...)
}
\arguments{
\item{...}{Additional arguments passed to \link{weighted.quantile}.}

\item{probability_w}{An object of class 'post_w', created with either the \link{exact_week_probability} or \link{week_probability}.}
}
\value{
Numeric vector of quantiles.
}
\description{
A wrapper for the \link{weighted.quantile} function of the \link{spatstat} package designed to work with 'post_w' objects.
}
