% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatteryPrognostics.R
\docType{package}
\name{BatteryPrognostics}
\alias{BatteryPrognostics}
\alias{BatteryPrognostics-package}
\title{Battery prognostics}
\description{
Filtering and prediction for state-of-charge and state-of-health prognostics for Li-ion batteries.
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
