% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCKLookupTable.R
\name{extractLookupTables}
\alias{extractLookupTables}
\title{Extract look-up tables.}
\usage{
extractLookupTables(current, voltage, time_s, t_k,
  capacityScalingParameters = NULL, h = 0.65, SOCList = seq(5, 95, 5),
  IList = c(-10, -7.5, -5, -2.5, -1.25, -0.25, 0.25, 1.25, 2.5, 5, 7.5, 10))
}
\arguments{
\item{current}{A periodic current profile.}

\item{voltage}{A periodic voltage profile.}

\item{time_s}{The sampling time.}

\item{t_k}{The time intervals for each of the RC-units.}

\item{h}{R0 reduction fraction.}

\item{SOCList}{A vector of possible 'SOC' values.}

\item{IList}{A vector of possible 'Current' values.}
}
\value{
A list of look-up table matrices.
}
\description{
Extracts RC2 parameter look-up tables from simple charge/discharge cycling profile.
}
